Fix #335: Invert checkbox/radiobutton
Create new assets for checkbox and radiobutton as svg and export as png.

All assets are in one SVG file, each on a separate layer. Sprites and
Backgrounds are referenced in these layers, so for example you can
change the checkmark and it is changed for all assets automatically.

The shellscript can be used to export one or more specific assets or all
at once.